+++ src/sass4clj/api.clj
@@ -10,11 +10,15 @@
            (.endsWith (.getName file) ".sass") )
        (not (.startsWith (.getName file) "_"))))
 
-(defn find-main-files [source-paths]
+(defn find-main-files [source-paths {:keys [inputs]}]
   (mapcat (fn [source-path]
             (let [file (io/file source-path)]
-              (->> (file-seq file)
-                   (filter main-file?)
+              (->> (if inputs
+                     (->> inputs
+                          (map #(io/file file %))
+                          (filter #(.exists %)))
+                     (->> (file-seq file)
+                          (filter main-file?)))
                    (map (fn [x] [(.getPath x) (.toString (.relativize (.toURI file) (.toURI x)))])))))
           source-paths))
 
@@ -48,6 +52,7 @@
         (throw e)))))
 
 (s/def ::source-paths (s/coll-of string? :into vec))
+(s/def ::inputs (s/coll-of string? :into vec))
 (s/def ::auto boolean?)
 (s/def ::help boolean?)
 (s/def ::target-path string?)
@@ -55,7 +60,7 @@
 (s/def ::verbosity #{1 2})
 (s/def ::output-style #{:nested :compact :expanded :compressed})
 (s/def ::options (s/keys :req-un [::source-paths ::target-path]
-                         :opt-un [::auto ::help ::source-map ::verbosity ::output-style]))
+                         :opt-un [::inputs ::auto ::help ::source-map ::verbosity ::output-style]))
 
 (defn build [{:keys [source-paths auto] :as options}]
   (when-not (s/valid? ::options options)
@@ -63,9 +68,9 @@
   (let [options (dissoc options :source-paths)]
     (if auto
       (watcher/start source-paths (fn [& _]
-                                    (let [main-files (find-main-files source-paths)]
+                                    (let [main-files (find-main-files source-paths options)]
                                       (compile-sass main-files options))))
-      (let [main-files (find-main-files source-paths)]
+      (let [main-files (find-main-files source-paths options)]
         (compile-sass main-files options)))))
 
 (defn start [options]